(WIP) install and run molecule on deployment host
This commit adds `requirements.txt` and `constraints.txt` files to
define the python packages required for running molecule. These will be
used in the process of constructing a python virtualenv specifically for
this purpose.

A deploy script has been added that is intended to be run on an Ubuntu
system that has a deployment of openstack-ansible. The script will
create the virtualenv, install the python packages, clone this repo,
checkout the specified branch and run `molecule converge` and `molecule
verify` on the system.