(ASC-572) (ASC-613) (ASC-614) using utility container for cinder
Prior to this PR, the tests had been executed on Master(Queens), Pike, and Pike-RC where openstack cli is on nova compute nodes (tested with compute1). Several tests are failling when the same system tests are running against Newton because openstack cli is not installed on compute1 in Newton.

This PR will use the openstack cli on utility container in infra1 node. Using openstack cli on the utility container is recommended by OpenStack.