ASC-636 Fix bootable zone key
This commit updates the data key supplied to the
`create_bootable_volume` helper in the `test_create_bootable_volume.py`
test. The data key used in the helper was changed from
`availability_zone` to `zone`. This change simply updates the calling
code to pass in a key of the same name.